apps/web/backend/requirements.txt
# Vibe Analytics Studio Backend Dependencies

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.3

# File Processing
openpyxl>=3.1.2
python-multipart>=0.0.6
werkzeug>=3.0.0

# Environment Management
python-dotenv>=1.0.0

# Development Dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Enhanced CSV Processing
chardet>=5.0.0
python-dateutil>=2.8.0
regex>=2023.0.0
